requirements.txt
google-genai==1.28.0
numpy
pandas